General fixes for Half Life + Mesh cap hack

+++ glRemixRenderer/structs.h
@@ -13,7 +13,7 @@ namespace glRemix
 
 struct MeshRecord
 {
-    UINT32 mesh_id;
+    UINT64 mesh_id;
 
     UINT32 blas_vb_ib_idx;
     UINT32 mv_idx;  // index into model view array
@@ -26,7 +26,7 @@ struct MeshRecord
     XMFLOAT3 max_bb;
 
     // For garbage collection, last frame this mesh record was accessed
-    UINT32 last_frame;
+    UINT64 last_frame;
 
     bool visible = true;  // can be toggled in the imgui
 };
